Changed

+++ 2.2dars/2.2dars.Api/Services/StudentService.cs
@@ -10,7 +10,11 @@ public class StudentService
     public StudentService()
     {
         studentFilePath = "../../../Data/Students.json";
-        File.WriteAllText(studentFilePath, "[]");
+
+        if(File.Exists(studentFilePath) is false)
+        {
+            File.WriteAllText(studentFilePath, "[]");
+        }
     }
 
     public Student AddStudent(Student student)
